refactor(linter): Move shellcheck logic to new file + finalize refactor
Make sure `__init__.py` only re-exports the actual task methods, not any helper functions.

Checked that the task list was still identical by `diff`-ing the output of `dda inv -l` before and after the refactor
